accesswash_platform/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.1.9
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==24.3
django-leaflet==0.32.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
